Rename file, spelling error
