+++ src/LogExpert/LogExpert.csproj
@@ -103,7 +103,9 @@
     <Compile Include="Controls\LogWindow\LogWindowsPublic.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Dialogs\LogCellEditingControl.cs" />
+    <Compile Include="Dialogs\LogCellEditingControl.cs">
+      <SubType>Component</SubType>
+    </Compile>
     <Compile Include="Dialogs\LogGridCell.cs" />
     <Compile Include="Dialogs\LogTextColumn.cs" />
     <Compile Include="Extensions\ComboBoxExtensions.cs" />
@@ -303,11 +305,11 @@
     <Compile Include="Dialogs\GotoLineDialog.Designer.cs">
       <DependentUpon>GotoLineDialog.cs</DependentUpon>
     </Compile>
-    <Compile Include="Dialogs\HilightDialog.cs">
+    <Compile Include="Dialogs\HighlightDialog.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Dialogs\HilightDialog.Designer.cs">
-      <DependentUpon>HilightDialog.cs</DependentUpon>
+    <Compile Include="Dialogs\HighlightDialog.Designer.cs">
+      <DependentUpon>HighlightDialog.cs</DependentUpon>
     </Compile>
     <Compile Include="Dialogs\RegexHelperDialog.cs">
       <SubType>Form</SubType>
@@ -394,8 +396,8 @@
       <DependentUpon>GotoLineDialog.cs</DependentUpon>
       <SubType>Designer</SubType>
     </EmbeddedResource>
-    <EmbeddedResource Include="Dialogs\HilightDialog.resx">
-      <DependentUpon>HilightDialog.cs</DependentUpon>
+    <EmbeddedResource Include="Dialogs\HighlightDialog.resx">
+      <DependentUpon>HighlightDialog.cs</DependentUpon>
       <SubType>Designer</SubType>
     </EmbeddedResource>
     <EmbeddedResource Include="Dialogs\RegexHelperDialog.resx">
